agregadas vistas para creacion, eliminacion y modificacion de articulos de una tienda

+++ codigo/controllers/TiendasController.php
@@ -209,119 +209,143 @@ class TiendasController extends Controller
      * @param array $DatosArticulos
      * @return \yii\web\Response
      */
-    public function actionCrearArticulo($Tienda_id, $DatosArticulos)
+    public function actionCrearArticulo($Tienda_id)
     {
-        $categoria = Categorias::findOne(['id' => $DatosArticulos['categoria_id']]); 
-        $etiqueta =  Etiquetas::findOne( ['id' => $DatosArticulos['etiqueta_id']]);
-        // Busca si el artículo ya existe en los artículos comunes
-        $comun = Articulo::findOne(['id' => $DatosArticulos['id'], 'tipo_marcado' => 'comun']);
-        if ($comun) {
-            // Si el artículo común existe, crea una relación con la tienda
-            $ArticuloTienda = new ArticulosTienda();
-            $historico = new Historico();
-            $ArticuloTienda->tienda_id = $Tienda_id;
-            $ArticuloTienda->articulo_id = $comun->id;
-            $ArticuloTienda->precio = $DatosArticulos['precio'];
-            $historico->tienda_id = $Tienda_id;
-            $historico->articulo_id = $comun->id;
-            $historico->precio = $DatosArticulos['precio'];
-            $historico->fecha = date('Y-m-d H:i:s');
-            $historico->save();
-            
-        } else {
-            // Si el artículo no existe, crea un nuevo artículo específico para la tienda
-            $Articulo = new Articulo();
-            $registro = new RegistroUsuarios();
-            $Articulo->nombre = $DatosArticulos['nombre'];
-            $Articulo->descripcion = $DatosArticulos['descripcion'];
-            $Articulo->categoria_id = $categoria->id;
-            $Articulo->etiqueta_id = $etiqueta->id;
-            $Articulo->imagen_ppal = $DatosArticulos['imagen_ppal'];
-            $Articulo->visible = 1;
-            $Articulo->cerrado = 0;
-            $Articulo->tipo_marcado = 'particular'; // Es un artículo particular de la tienda
-            $Articulo->registro_id = Yii::$app->user->id;
-            $registro->fecha_creacion = date('Y-m-d H:i:s');
-            $registro->creador_id = Yii::$app->user->id;
-            $registro->notas_admin = 'Artículo creado por el usuario';
-            $registro->save();
-            
-
-            if ($Articulo->save()) {
+        $model = new Articulo();
+        $categorias = Categorias::find()->all();
+        $etiquetas = Etiquetas::find()->all();
+    
+        if ($model->load(Yii::$app->request->post())) {
+            $DatosArticulos = Yii::$app->request->post('Articulo');
+            $categoria = Categorias::findOne(['id' => $DatosArticulos['categoria_id']]); 
+            $etiqueta = Etiquetas::findOne(['id' => $DatosArticulos['etiqueta_id']]);
+            // Busca si el artículo ya existe en los artículos comunes
+            $comun = Articulo::findOne(['id' => $DatosArticulos['id'], 'tipo_marcado' => 'comun']);
+            if ($comun) {
+                // Si el artículo común existe, crea una relación con la tienda
                 $ArticuloTienda = new ArticulosTienda();
                 $historico = new Historico();
                 $ArticuloTienda->tienda_id = $Tienda_id;
-                $ArticuloTienda->articulo_id = $Articulo->id;
+                $ArticuloTienda->articulo_id = $comun->id;
                 $ArticuloTienda->precio = $DatosArticulos['precio'];
-                $Articulo->Articulo_tienda_id= $ArticuloTienda->id;
                 $historico->tienda_id = $Tienda_id;
-                $historico->articulo_id = $Articulo->id;
+                $historico->articulo_id = $comun->id;
                 $historico->precio = $DatosArticulos['precio'];
                 $historico->fecha = date('Y-m-d H:i:s');
-                $Articulo->save();
                 $historico->save();
             } else {
-                Yii::$app->session->setFlash('error', 'Ha habido un error al crear el artículo.');
-                return $this->redirect(['view-store', 'id' => $Tienda_id]);
+                // Si el artículo no existe, crea un nuevo artículo específico para la tienda
+                $Articulo = new Articulo();
+                $registro = new RegistroUsuarios();
+                $Articulo->nombre = $DatosArticulos['nombre'];
+                $Articulo->descripcion = $DatosArticulos['descripcion'];
+                $Articulo->categoria_id = $categoria->id;
+                $Articulo->etiqueta_id = $etiqueta->id;
+                $Articulo->imagen_ppal = $DatosArticulos['imagen_ppal'];
+                $Articulo->visible = 1;
+                $Articulo->cerrado = 0;
+                $Articulo->tipo_marcado = 'particular'; // Es un artículo particular de la tienda
+                $Articulo->registro_id = Yii::$app->user->id;
+                $registro->fecha_creacion = date('Y-m-d H:i:s');
+                $registro->creador_id = Yii::$app->user->id;
+                $registro->notas_admin = 'Artículo creado por el usuario';
+                $registro->save();
+    
+                if ($Articulo->save()) {
+                    $ArticuloTienda = new ArticulosTienda();
+                    $historico = new Historico();
+                    $ArticuloTienda->tienda_id = $Tienda_id;
+                    $ArticuloTienda->articulo_id = $Articulo->id;
+                    $ArticuloTienda->precio = $DatosArticulos['precio'];
+                    $Articulo->Articulo_tienda_id = $ArticuloTienda->id;
+                    $historico->tienda_id = $Tienda_id;
+                    $historico->articulo_id = $Articulo->id;
+                    $historico->precio = $DatosArticulos['precio'];
+                    $historico->fecha = date('Y-m-d H:i:s');
+                    $Articulo->save();
+                    $historico->save();
+                } else {
+                    Yii::$app->session->setFlash('error', 'Ha habido un error al crear el artículo.');
+                    return $this->redirect(['view-store', 'id' => $Tienda_id]);
+                }
             }
+    
+            // Guarda la relación del artículo con la tienda
+            if ($ArticuloTienda->save()) {
+                Yii::$app->session->setFlash('success', 'Artículo creado o vinculado con éxito.');
+            } else {
+                Yii::$app->session->setFlash('error', 'Ha habido un error.');
+            }
+    
+            return $this->redirect(['view-store', 'id' => $Tienda_id]);
         }
-
-        // Guarda la relación del artículo con la tienda
-        if ($ArticuloTienda->save()) {
-            Yii::$app->session->setFlash('success', 'Artículo creado o vinculado con éxito.');
-        } else {
-            Yii::$app->session->setFlash('error', 'Ha habido un error.');
-        }
-
-        return $this->redirect(['view-store', 'id' => $Tienda_id]);
+    
+        return $this->render('crear-articulo', [
+            'model' => $model,
+            'categorias' => $categorias,
+            'etiquetas' => $etiquetas,
+        ]);
     }
 
-    public function actionModificarArticulo($Tienda_id, $Articulo_id, $DatosArticulos)
+    public function actionModificarArticulo($Tienda_id, $Articulo_id)
     {
         // Busca la relación del artículo con la tienda
         $ArticuloTienda = ArticulosTienda::findOne(['tienda_id' => $Tienda_id, 'articulo_id' => $Articulo_id]);
         if (!$ArticuloTienda) {
             throw new NotFoundHttpException('El artículo no existe en esta tienda.');
         }
-        $categoria = Categorias::findOne( $DatosArticulos['categoria_id']); 
-        $etiqueta =  Etiquetas::findOne( $DatosArticulos['etiqueta_id']);
-
-        // Verifica si el artículo es común
-        $Articulo = Articulo::findOne($Articulo_id);
-        $historico = new Historico();
-        if ($Articulo->tipo_marcado == 'comun') {
-            // Solo permite la modificación del precio para artículos comunes
-            $ArticuloTienda->precio = $DatosArticulos['precio'];
-            $historico->tienda_id = $Tienda_id;
-            $historico->articulo_id = $Articulo->id;
-            $historico->precio = $ArticuloTienda->precio;
-            $historico->fecha = date('Y-m-d H:i:s');
-            $historico->save();
-
-        } else {
-            // Permite la modificación completa para artículos específicos de la tienda
-            $Articulo->nombre = $DatosArticulos['nombre'];
-            $Articulo->descripcion = $DatosArticulos['descripcion'];
-            $Articulo->categoria_id = $categoria->id;
-            $Articulo->etiqueta_id = $etiqueta->id;
-            $Articulo->imagen_ppal = $DatosArticulos['imagen_ppal'];
-
-            if (!$Articulo->save()) {
-                Yii::$app->session->setFlash('error', 'Ha habido un error al modificar el artículo.');
-                return $this->redirect(['view-store', 'id' => $Tienda_id]);
+    
+        $model = Articulo::findOne($Articulo_id);
+        $categorias = Categorias::find()->all();
+        $categorias = ArrayHelper::map($categorias,'id','nombre');
+        $etiquetas = Etiquetas::find()->all();
+        $etiquetas = ArrayHelper::map($etiquetas,'id','nombre');
+    
+        if ($model->load(\Yii::$app->request->post())) {
+            $DatosArticulos = \Yii::$app->request->post('Articulo');
+            $categoria = Categorias::findOne(['id' => $DatosArticulos['categoria_id']]); 
+            $etiqueta = Etiquetas::findOne(['id' => $DatosArticulos['etiqueta_id']]);
+            $historico = new Historico();
+    
+            // Verifica si el artículo es común
+            if ($model->tipo_marcado == 'comun') {
+                // Solo permite la modificación del precio para artículos comunes
+                $ArticuloTienda->precio = $DatosArticulos['precio'];
+                $historico->tienda_id = $Tienda_id;
+                $historico->articulo_id = $model->id;
+                $historico->precio = $ArticuloTienda->precio;
+                $historico->fecha = date('Y-m-d H:i:s');
+                $historico->save();
+            } else {
+                // Permite la modificación completa para artículos específicos de la tienda
+                $model->nombre = $DatosArticulos['nombre'];
+                $model->descripcion = $DatosArticulos['descripcion'];
+                $model->categoria_id = $categoria->id;
+                $model->etiqueta_id = $etiqueta->id;
+                $model->imagen_ppal = $DatosArticulos['imagen_ppal'];
+    
+                if (!$model->save()) {
+                    Yii::$app->session->setFlash('error', 'Ha habido un error al modificar el artículo.');
+                    return $this->redirect(['view-store', 'id' => $Tienda_id]);
+                }
             }
+    
+            // Guarda los cambios en la relación del artículo con la tienda
+            if ($ArticuloTienda->save()) {
+                Yii::$app->session->setFlash('success', 'Artículo modificado con éxito.');
+            } else {
+                Yii::$app->session->setFlash('error', 'Ha habido un error.');
+            }
+    
+            return $this->redirect(['view-store', 'id' => $Tienda_id]);
         }
-
-        // Guarda los cambios en la relación del artículo con la tienda
-        if ($ArticuloTienda->save()) {
-            Yii::$app->session->setFlash('success', 'Artículo modificado con éxito.');
-        } else {
-            Yii::$app->session->setFlash('error', 'Ha habido un error.');
-        }
-
-        return $this->redirect(['view-store', 'id' => $Tienda_id]);
+    
+        return $this->render('modificar-articulo', [
+            'model' => $model,
+            'categorias' => $categorias,
+            'etiquetas' => $etiquetas,
+        ]);
     }
-
     /**
      * Deletes or unlinks an article from the store.
      * If the article has price history, it will be hidden instead of deleted.
@@ -337,43 +361,47 @@ class TiendasController extends Controller
         if (!$ArticuloTienda) {
             throw new NotFoundHttpException('El artículo no existe en esta tienda.');
         }
-
-        $historico = Historico::find()->where(['tienda_id' => $Tienda_id, 'articulo_id' => $Articulo_id])->all();
-
-        if(!$historico){
-            throw new NotFoundHttpException('El artículo no tiene historico de precios para esta tienda.');
-        }
-        // Verifica si el artículo tiene histórico de precios
-        if ($historico) {
-            $ArticuloTienda->visible = 0; // Oculta el artículo
-            $message = 'Artículo ocultado con éxito.';
-        } else {
-            // Verifica si el artículo es común
-            $Articulo = Articulo::findOne($Articulo_id);
-            if ($Articulo->tipo_marcado == 'comun') {
-                // Desvincula el artículo común de la tienda
-                $ArticuloTienda->delete();
-                $message = 'Artículo desvinculado con éxito.';
-            } else {
-                // Elimina el artículo particular de la tienda
-                $ArticuloTienda->delete();
-                $Articulo->delete();
-                $Articulo->save();
-                $message = 'Artículo eliminado con éxito.';
+    
+        if (Yii::$app->request->post()) {
+            $historico = Historico::find()->where(['tienda_id' => $Tienda_id, 'articulo_id' => $Articulo_id])->all();
+    
+            if (!$historico) {
+                throw new NotFoundHttpException('El artículo no tiene historico de precios para esta tienda.');
             }
+    
+            // Verifica si el artículo tiene histórico de precios
+            if ($historico) {
+                $ArticuloTienda->visible = 0; // Oculta el artículo
+                $message = 'Artículo ocultado con éxito.';
+            } else {
+                // Verifica si el artículo es común
+                $Articulo = Articulo::findOne($Articulo_id);
+                if ($Articulo->tipo_marcado == 'comun') {
+                    // Desvincula el artículo común de la tienda
+                    $ArticuloTienda->delete();
+                    $message = 'Artículo desvinculado con éxito.';
+                } else {
+                    // Elimina el artículo particular de la tienda
+                    $ArticuloTienda->delete();
+                    $Articulo->delete();
+                    $message = 'Artículo eliminado con éxito.';
+                }
+            }
+    
+            // Guarda los cambios
+            if ($ArticuloTienda->save()) {
+                Yii::$app->session->setFlash('success', $message);
+            } else {
+                Yii::$app->session->setFlash('error', 'Ha habido un error.');
+            }
+    
+            return $this->redirect(['view-store', 'id' => $Tienda_id]);
         }
-
-        // Guarda los cambios
-        if ($ArticuloTienda->save()) {
-            Yii::$app->session->setFlash('success', $message);
-        } else {
-            Yii::$app->session->setFlash('error', 'Ha habido un error.');
-        }
-
-        return $this->redirect(['view-store', 'id' => $Tienda_id]);
-
+    
+        return $this->render('eliminar-articulo', [
+            'model' => $ArticuloTienda,
+        ]);
     }
-
     /**
      * Displays the price history of a specific article in a store.
      * @param int $Tienda_id
